Fix flickering question form while animated
MixItUp set position: absolute for all items in a list while it's animating, making the bottom margin ignored for elements outside of the list. So we add a bottom margin for the list itself (which is not absolute positioned) and remove the margin of the last item so it can't interfere with the layout outside the list.